Added GA, split Blog Styles
