+++ src/pages/{MarkdownRemark.frontmatter__slug}.js
@@ -1,6 +1,9 @@
 import React from "react"
-import { graphql, navigate } from "gatsby"
-import styled from "styled-components"
+import {
+  graphql,
+  // navigate
+} from "gatsby"
+// import styled from "styled-components"
 
 export default function Template({
   data, // this prop will be injected by the GraphQL query below.
@@ -8,12 +11,11 @@ export default function Template({
   const { markdownRemark } = data // data.markdownRemark holds your post data
   const { frontmatter, html } = markdownRemark
   return (
-    <div className="blog-post-container">
-      <div className="blog-post">
+    <div>
+      <div >
         <h1>{frontmatter.title}</h1>
-        <h2>{frontmatter.date}</h2>
+        <h2>Date: {frontmatter.date}</h2>
         <div
-          className="blog-post-content"
           dangerouslySetInnerHTML={{ __html: html }}
         />
       </div>
